Refactor InjectBackendClient into separate module
again, for readability (and potential re-use).
